backend/requirements.txt
# Python dependencies for GitHub Webhook Backend
# Flask web framework and extensions
flask==3.0.0
flask-cors==4.0.0

# MongoDB driver
pymongo==4.6.1

# Environment variable management
python-dotenv==1.0.0

# Production WSGI server
gunicorn==21.2.0
